my additions
Nothing dramatic here.  Added Berago's CSS, and added a reference to
the bootstrap responsive
style sheet so that we can respond to different window sizes.

Moved the navbar to the bottom, simplified things so that the first
thing you see is the search bar, then glasses, then the table of
results.
Right now we can display fixed percentages on the glasses.  It is not
yet changing the numbers on the glasses depending on what twitter ID
you enter, but I didn't want to dig too deep into that since we're
going to change that when we get the details on the API.

Questions to address before going further:

We need to make the image responsive so that it displays on different
window sizes - my default on how to do this would be to simply create
4-5 different sizes depending on the width of the screen and to set
minimum width criteria in the CSS.  My concern is making sure the text
lines up on top of the image when we resize - I assume there has to be
 but if we have to manually
align for each image I guess we can do that.

Otherwise, aside from tweaking/ui fine tuning and some potential
illustrator work for the graphics i think we're on our way...